airflow_ml_dags/images/airflow-pipeline/requirements.txt
numpy==1.20.1
pandas==1.2.4
scikit-learn==0.24.2
joblib==0.17.0
click==7.1.2
pytest==6.2.3
hydra_core==1.0.6
omegaconf==2.0.6
dataclasses==0.6
git+https://github.com/made-ml-in-prod-2021/cherninkiy.git#subdirectory=ml_project&egg=ml-pipeline

